+++ src/factories/spell-factory.ts
@@ -1,21 +1,21 @@
+import { ServiceKeys, ServiceLocator } from '@/components/core/service-locator';
 import { FireBall } from '@/objects/spells/direct-spells/fire-ball';
 import { Blink } from '@/objects/spells/effect-spells/blink';
 import { SPELLS } from '@/utils/constants';
 import { AnimationKeys } from '@/utils/animation-keys';
 
 export class SpellFactory {
-  spellGroup: Phaser.Physics.Arcade.Group;
+  private spellGroup: Phaser.Physics.Arcade.Group;
 
   constructor(private scene: Phaser.Scene) {
-    this.scene = scene;
-
     this.spellGroup = this.scene.physics.add.group({
       runChildUpdate: true,
       allowGravity: false,
     });
   }
 
-  createFireball(x: number, y: number, direction: number): void {
+  public createFireball(x: number, y: number, direction: number): FireBall {
+    const sandbox = ServiceLocator.resolve(ServiceKeys.sandbox);
     const offsetX = direction * 20;
 
     const fireBall = new FireBall({
@@ -23,6 +23,7 @@ export class SpellFactory {
       position: { x: x + offsetX, y: y },
       keyName: SPELLS.FIREBALL,
       frame: 0,
+      sandbox,
       animation: {
         main: AnimationKeys.Spells.Fireball.Main,
         destroy: AnimationKeys.Spells.Fireball.Destroy,
@@ -33,21 +34,31 @@ export class SpellFactory {
     });
 
     this.spellGroup.add(fireBall, true);
+
+    return fireBall;
   }
 
-  createBlink(x: number, y: number, direction: number): void {
+  public createBlink(x: number, y: number, direction: number): Blink {
+    const sandbox = ServiceLocator.resolve(ServiceKeys.sandbox);
+
     const blink = new Blink({
       scene: this.scene,
       position: { x: x - 4 * direction, y: y + 5 },
       keyName: SPELLS.BLINK,
       frame: 0,
+      sandbox,
       animation: {
         main: AnimationKeys.Spells.Blink.Main,
       },
+      direction: direction,
     });
+
+    this.spellGroup.add(blink, true);
+
+    return blink;
   }
 
-  getSpells(): Phaser.Physics.Arcade.Group {
+  public getSpells(): Phaser.Physics.Arcade.Group {
     return this.spellGroup;
   }
 }